feat: producto

+++ src/main/java/com/cibertec/desappwebt2/infrastucture/database/repositories/ProductoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductoRepository extends CrudRepository<ProductoEntity, Long> {
+public interface ProductoRepository extends CrudRepository<ProductoEntity, Integer> {
 }
